Se movio test de MovieServiceTest a MovieControllerTest
