+++ PAC.Vidly.WebApi.UnitTests/Controllers/MovieControllerTests.cs
@@ -6,6 +6,7 @@ using PAC.Vidly.WebApi.Controllers.Movies.Models;
 using PAC.Vidly.WebApi.DataAccess;
 using PAC.Vidly.WebApi.Services.Movies;
 using PAC.Vidly.WebApi.Services.Movies.Entities;
+using PAC.Vidly.WebApi.Services.Users.Entities;
 
 namespace PAC.Vidly.WebApi.UnitTests.Controllers
 {
@@ -45,11 +46,9 @@ namespace PAC.Vidly.WebApi.UnitTests.Controllers
             var service = new MovieService(repositoryMock.Object);
             var controller = new MovieController(service);
 
-            var request = new Movie
+            var request = new CreateMovieRequest()
             {
-                Id = "test",
                 Name = string.Empty,
-                CreatorId = "test",
             };
 
             try
@@ -58,9 +57,29 @@ namespace PAC.Vidly.WebApi.UnitTests.Controllers
             }
             catch (Exception ex)
             {
-                ex.Message.Should().Be("Name cannot be empty");
+                ex.Message.Should().Be("Name cannot be empty or null");
             }
         }
+        
+        [TestMethod]
+        [ExpectedException(typeof(ArgumentNullException))]
+        public void Create_WhenNameIsNull_ShouldThrowException()
+        {
+            var request = new CreateMovieRequest()
+            {
+                Name = null,
+            };
+
+            try
+            {
+                _controller.Create(request);
+            }
+            catch (Exception ex)
+            {
+                ex.Message.Should().Be("Name cannot be empty or null");
+            }
+        }
+        
         #endregion
 
         #region GetAll
